+++ src/js/integrations/exports.ts
@@ -14,6 +14,7 @@ export { expoContextIntegration } from './expocontext';
 export { spotlightIntegration } from './spotlight';
 export { mobileReplayIntegration } from '../replay/mobilereplay';
 export { appStartIntegration } from '../tracing/integrations/appStart';
+export { nativeFramesIntegration } from '../tracing/integrations/nativeFrames';
 
 export {
   breadcrumbsIntegration,